+++ astropy/io/misc/asdf/tags/transform/basic.py
@@ -4,7 +4,7 @@
 from asdf import tagged
 from asdf.versioning import AsdfVersion
 
-from astropy.modeling import models, mappings
+from astropy.modeling import core, models, mappings
 from astropy.utils import minversion
 from astropy.modeling import functional_models
 from astropy.modeling.core import Model, CompoundModel
@@ -48,6 +48,12 @@ class TransformType(AstropyAsdfType):
                 param_and_model_constraints[constraint] = node[constraint]
         model._initialize_constraints(param_and_model_constraints)
 
+        if "input_units_equivalencies" in node:
+            # this still writes eqs. for compound, but operates on each sub model
+            if not isinstance(model, core.CompoundModel):
+                eqs = node['input_units_equivalencies']
+                model.input_units_equivalencies = eqs
+
         return model
 
     @classmethod
@@ -63,6 +69,7 @@ class TransformType(AstropyAsdfType):
 
     @classmethod
     def _to_tree_base_transform_members(cls, model, node, ctx):
+
         if getattr(model, '_user_inverse', None) is not None:
             node['inverse'] = model._user_inverse
 
@@ -92,6 +99,15 @@ class TransformType(AstropyAsdfType):
             bounds_nondefaults = {k:b for k, b in model.bounds.items() if any(b)}
             if bounds_nondefaults:
                 node['bounds'] = bounds_nondefaults
+
+        if model.input_units is not None:
+            if model.input_units_equivalencies:
+                input_unit_equivalencies = {}
+                for in_unit in model.input_units:
+                    eq = model.input_units_equivalencies[in_unit]
+                    unit_equiv = eq
+                    input_unit_equivalencies[in_unit] = unit_equiv
+                node['input_units_equivalencies'] = input_unit_equivalencies
 
     @classmethod
     def to_tree_transform(cls, model, ctx):
@@ -151,6 +167,8 @@ class ConstantType(TransformType):
             return functional_models.Const1D(node['value'])
         elif node['dimensions'] == 2:
             return functional_models.Const2D(node['value'])
+        else:
+            raise TypeError('Only 1D and 2D constant models are supported.')
 
     @classmethod
     def to_tree_transform(cls, data, ctx):